Update README.md
Just corrected an indentation error that didn't make a heck of a difference, in fact, I think it wasn't even showing at all. But anyway, I'm a stickler for proper formatting. I hate the way my nested lists are showing. The letters like "a" and "b" aren't indenting. I don't know how to make the Markdown work with nested lists that have BOTH NUMBERS AND LETTERS!